+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Working Log
 
-### 1.0.0 - 2021-12-09
+### 1.0.1 - 2021-12-21
+
+#### Added
+-   flashing logic. 
+
+#### Updated
+
+-   selection logic for getting nearest feature.
+-   sample data.
+
+
+### 1.0.0 - 2021-12-11
 
 #### Added